moved testData.jsx to new mock folder and updated references
