+++ spec/src/Overview.test.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import Overview from '../../client/src/components/Overview/Overview.jsx';
-import testData from '../../client/src/testData.jsx';
+import testData from '../mock/testData.jsx';
 
 describe('Render Overview Module', () => {
   const setup = (current, currentStyles, stars = 5, favorites = [], reviewsQty = 5) => {
